+++ bittytax/conv/output_excel.py
@@ -103,13 +103,15 @@ class Worksheet(object):
                 TransactionOutRecord.TYPE_INTEREST,
                 TransactionOutRecord.TYPE_DIVIDEND,
                 TransactionOutRecord.TYPE_INCOME,
-                TransactionOutRecord.TYPE_GIFT_RECEIVED)
+                TransactionOutRecord.TYPE_GIFT_RECEIVED,
+                TransactionOutRecord.TYPE_AIRDROP)
 
     SELL_LIST = (TransactionOutRecord.TYPE_WITHDRAWAL,
                  TransactionOutRecord.TYPE_SPEND,
                  TransactionOutRecord.TYPE_GIFT_SENT,
                  TransactionOutRecord.TYPE_GIFT_SPOUSE,
-                 TransactionOutRecord.TYPE_CHARITY_SENT)
+                 TransactionOutRecord.TYPE_CHARITY_SENT,
+                 TransactionOutRecord.TYPE_LOST)
     SHEETNAME_MAX_LEN = 31
     MAX_COL_WIDTH = 30