+++ bittytax/import_records.py
@@ -153,19 +153,22 @@ class TransactionRow(object):
     OPT = 'Optional'
     MAN = 'Mandatory'
 
-    TYPE_VALIDATION = {TR.TYPE_DEPOSIT:       [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
-                       TR.TYPE_MINING:        [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
-                       TR.TYPE_STAKING:       [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
-                       TR.TYPE_INTEREST:      [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
-                       TR.TYPE_DIVIDEND:      [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
-                       TR.TYPE_INCOME:        [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
-                       TR.TYPE_GIFT_RECEIVED: [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
-                       TR.TYPE_WITHDRAWAL:    [MAN, None, None, None, MAN, MAN, OPT, OPT, OPT, OPT],
-                       TR.TYPE_SPEND:         [MAN, None, None, None, MAN, MAN, OPT, OPT, OPT, OPT],
-                       TR.TYPE_GIFT_SENT:     [MAN, None, None, None, MAN, MAN, OPT, OPT, OPT, OPT],
-                       TR.TYPE_GIFT_SPOUSE:   [MAN, None, None, None, MAN, MAN, OPT, OPT, OPT, OPT],
-                       TR.TYPE_CHARITY_SENT:  [MAN, None, None, None, MAN, MAN, OPT, OPT, OPT, OPT],
-                       TR.TYPE_TRADE:         [MAN, MAN, MAN, OPT, MAN, MAN, OPT, OPT, OPT, OPT]}
+    TYPE_VALIDATION = {
+            TR.TYPE_DEPOSIT:       [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
+            TR.TYPE_MINING:        [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
+            TR.TYPE_STAKING:       [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
+            TR.TYPE_INTEREST:      [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
+            TR.TYPE_DIVIDEND:      [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
+            TR.TYPE_INCOME:        [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
+            TR.TYPE_GIFT_RECEIVED: [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
+            TR.TYPE_AIRDROP:       [MAN, MAN, MAN, OPT, None, None, None, OPT, OPT, OPT],
+            TR.TYPE_WITHDRAWAL:    [MAN, None, None, None, MAN, MAN, OPT, OPT, OPT, OPT],
+            TR.TYPE_SPEND:         [MAN, None, None, None, MAN, MAN, OPT, OPT, OPT, OPT],
+            TR.TYPE_GIFT_SENT:     [MAN, None, None, None, MAN, MAN, OPT, OPT, OPT, OPT],
+            TR.TYPE_GIFT_SPOUSE:   [MAN, None, None, None, MAN, MAN, OPT, OPT, OPT, OPT],
+            TR.TYPE_CHARITY_SENT:  [MAN, None, None, None, MAN, MAN, OPT, OPT, OPT, OPT],
+            TR.TYPE_LOST:          [MAN, None, None, None, MAN, MAN, OPT, None, None, None],
+            TR.TYPE_TRADE:         [MAN, MAN, MAN, OPT, MAN, MAN, OPT, OPT, OPT, OPT]}
 
     TRANSFER_TYPES = (TR.TYPE_DEPOSIT, TR.TYPE_WITHDRAWAL)
 
@@ -226,7 +229,16 @@ class TransactionRow(object):
         if buy_asset:
             buy = Buy(t_type, buy_quantity, buy_asset, buy_value)
         if sell_asset:
-            sell = Sell(t_type, sell_quantity, sell_asset, sell_value)
+            if t_type == TR.TYPE_LOST:
+                if sell_value is None:
+                    sell_value = Decimal(0)
+
+                sell = Sell(t_type, sell_quantity, sell_asset, sell_value)
+                if config.lost_buyback:
+                    buy = Buy(t_type, sell_quantity, sell_asset, sell_value)
+                    buy.acquisition = True
+            else:
+                sell = Sell(t_type, sell_quantity, sell_asset, sell_value)
         if fee_asset:
             # Fees are added as a separate spend transaction
             fee = Sell(TR.TYPE_SPEND, fee_quantity, fee_asset, fee_value)